# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transaction Approval API",
        "description": "API documentation for Transaction Approval",
        "version": "1.0.0"
    },
    "paths": {
        "/api/approval-stages": {
            "post": {
                "tags": [
                    "ApprovalStage"
                ],
                "summary": "Create approval stage",
                "operationId": "52989e6a14178a727bcc848e3acf8044",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "approver_id"
                                ],
                                "properties": {
                                    "approver_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Approval stage created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "approver_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-06-16T12:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2024-06-16T12:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/approval-stages/{id}": {
            "put": {
                "tags": [
                    "ApprovalStage"
                ],
                "summary": "Update approval stage",
                "operationId": "09e3ea8a8b87944defd865cd349851f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "approver_id"
                                ],
                                "properties": {
                                    "approver_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Approval stage updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "approver_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-06-16T12:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2024-06-16T12:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid approval stage ID"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/approvers": {
            "post": {
                "tags": [
                    "Approver"
                ],
                "summary": "Create new approver",
                "operationId": "8fddc9e935853028d340f48240c4f3ef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ana"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Approver created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Ana"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/expense": {
            "post": {
                "tags": [
                    "Expense"
                ],
                "summary": "Create new expense",
                "operationId": "a8e3ae7ca07f1c973a526d413dc78a7e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "integer",
                                        "example": 1000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Expense created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "amount": {
                                            "type": "integer",
                                            "example": 1000
                                        },
                                        "status_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-06-16T12:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2024-06-16T12:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/expense/{id}/approve": {
            "patch": {
                "tags": [
                    "Expense"
                ],
                "summary": "Approve expense",
                "operationId": "5ff2fd78df71a303e74d8a6c26c4dd0e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "approver_id"
                                ],
                                "properties": {
                                    "approver_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Expense approved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "amount": {
                                            "type": "integer",
                                            "example": 1000
                                        },
                                        "status_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-06-16T12:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2024-06-16T12:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Expense not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/expense/{id}": {
            "get": {
                "tags": [
                    "Expense"
                ],
                "summary": "Get expense by ID",
                "operationId": "051a76d052cc492fba7e7536cecfd1d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expense detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "amount": {
                                            "type": "integer",
                                            "example": 1000
                                        },
                                        "status_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-06-16T12:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2024-06-16T12:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Expense not found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "ApprovalStage",
            "description": "ApprovalStage"
        },
        {
            "name": "Approver",
            "description": "Approver"
        },
        {
            "name": "Expense",
            "description": "Expense"
        }
    ]
}